requirements.txt
celery==3.1.25
flower
django-bulma
django-cors-headers
django-debug-toolbar
django-extensions
django-filter
django-rest-framework
dj-stripe
pillow
pytest-django
python-decouple
social-auth-app-django
WeasyPrint
win10toast
